Moved all user credentials into a new Base test class for live tests to be able to share it amongst all test in the future
